change uint8 to float

+++ src/ops/accelerationconv.cpp
@@ -87,10 +87,10 @@ inline void CPUAccelerationConvOp::forward_compute() {
   auto input_tensor = getInputs()[0];
   auto output_tensor = getOutputs()[0];
 
-  uint8_t* input = reinterpret_cast<uint8_t*>(input_tensor->data_ptr);
-  uint8_t* output = reinterpret_cast<uint8_t*>(output_tensor->data_ptr);
-  uint8_t* weight = reinterpret_cast<uint8_t*>(weight_->data_ptr);
-  uint8_t* bias = bias_ ? reinterpret_cast<uint8_t*>(bias_->data_ptr) : nullptr;
+  float* input = reinterpret_cast<float*>(input_tensor->data_ptr);
+  float* output = reinterpret_cast<float*>(output_tensor->data_ptr);
+  float* weight = reinterpret_cast<float*>(weight_->data_ptr);
+  float* bias = bias_ ? reinterpret_cast<float*>(bias_->data_ptr) : nullptr;
 
   int ni = input_tensor->n_batch;
   int ci = input_tensor->channel;
@@ -109,7 +109,7 @@ inline void CPUAccelerationConvOp::forward_compute() {
   int ph = param_.ph;
   int pw = param_.pw;
 
-  uint8_t* temp_weight = nullptr;
+  float* temp_weight = nullptr;
   int x = 0, y = 0;
   // 卷积核的个数 = 输出的通道数
   int m = output_tensor->channel;
@@ -122,14 +122,14 @@ inline void CPUAccelerationConvOp::forward_compute() {
   /*循环batch中的每个输入*/
   for (int i = 0; i < ni; ++i) {
     /*用于存储经im2col转换后的输入特征矩阵*/
-    uint8_t* a = (uint8_t*)calloc(1024, sizeof(uint8_t));
+    float* a = (float*)calloc(1024, sizeof(float));
     tofree_.push_back(a);
 
     /*a是指向当前层所有卷积核的*/
-    uint8_t* b = weight;
+    float* b = weight;
     /*输出特征图个数*/
-    uint8_t* c = output + i * n * m;
-    uint8_t* im = input + i * ci * hi * wi;
+    float* c = output + i * n * m;
+    float* im = input + i * ci * hi * wi;
     /*如果是1*1的卷积，那么不用对输入特征进行转化*/
     if (kh * kw == 1) {
       a = im;
@@ -142,7 +142,7 @@ inline void CPUAccelerationConvOp::forward_compute() {
 }
 
 inline void CPUAccelerationConvOp::mm_generate(
-    uint8_t* matA, uint8_t* matB, uint8_t* matC, const int M, const int N,
+    float* matA, float* matB, float* matC, const int M, const int N,
     const int K, const int strideA, const int strideB, const int strideC) {
   //    printf("into mm_generate\n");
 
@@ -166,19 +166,19 @@ inline void CPUAccelerationConvOp::mm_generate(
 **        strideC     C的列数
 */
 inline void CPUAccelerationConvOp::coppersmith_winograd(
-    uint8_t* matA, uint8_t* matB, uint8_t* matC, int M, int N, int K,
+    float* matA, float* matB, float* matC, int M, int N, int K,
     int strideA, int strideB, int strideC) {
   // step 1:使用普通的矩阵
   if ((M <= 64) || (M % 2 != 0 || N % 2 != 0 || K % 2 != 0)) {
     return mm_generate(matA, matB, matC, M, N, K, strideA, strideB, strideC);
   }
-  // matC = calloc(M * strideC, sizeof(uint8_t));
+  // matC = calloc(M * strideC, sizeof(float));
   int offset = 0;
 
-  uint8_t* S1 =(uint8_t*) calloc((M / 2) * (K / 2), sizeof(uint8_t));
-  uint8_t* S2 =(uint8_t*) calloc((M / 2) * (K / 2), sizeof(uint8_t));
-  uint8_t* S3 =(uint8_t*) calloc((M / 2) * (K / 2), sizeof(uint8_t));
-  uint8_t* S4 =(uint8_t*) calloc((M / 2) * (K / 2), sizeof(uint8_t));
+  float* S1 =(float*) calloc((M / 2) * (K / 2), sizeof(float));
+  float* S2 =(float*) calloc((M / 2) * (K / 2), sizeof(float));
+  float* S3 =(float*) calloc((M / 2) * (K / 2), sizeof(float));
+  float* S4 =(float*) calloc((M / 2) * (K / 2), sizeof(float));
   tofree_.push_back(S1);
   tofree_.push_back(S2);
   tofree_.push_back(S3);
@@ -197,10 +197,10 @@ inline void CPUAccelerationConvOp::coppersmith_winograd(
       S4[idx] = matA[i * strideA + j + K / 2] - S2[idx];
     }
   }
-  uint8_t* T1 =(uint8_t*) calloc((K / 2) * (N / 2), sizeof(uint8_t));
-  uint8_t* T2 =(uint8_t*) calloc((K / 2) * (N / 2), sizeof(uint8_t));
-  uint8_t* T3 =(uint8_t*) calloc((K / 2) * (N / 2), sizeof(uint8_t));
-  uint8_t* T4 =(uint8_t*) calloc((K / 2) * (N / 2), sizeof(uint8_t));
+  float* T1 =(float*) calloc((K / 2) * (N / 2), sizeof(float));
+  float* T2 =(float*) calloc((K / 2) * (N / 2), sizeof(float));
+  float* T3 =(float*) calloc((K / 2) * (N / 2), sizeof(float));
+  float* T4 =(float*) calloc((K / 2) * (N / 2), sizeof(float));
   tofree_.push_back(T1);
   tofree_.push_back(T2);
   tofree_.push_back(T3);
@@ -221,7 +221,7 @@ inline void CPUAccelerationConvOp::coppersmith_winograd(
   }
 
   // M1 = A11*B11
-  uint8_t* M1 = (uint8_t*)calloc((M / 2) * (N / 2), sizeof(uint8_t));
+  float* M1 = (float*)calloc((M / 2) * (N / 2), sizeof(float));
   tofree_.push_back(M1);
   {
     printf("M1\n");
@@ -230,7 +230,7 @@ inline void CPUAccelerationConvOp::coppersmith_winograd(
   }
 
   // M2 = A12*B21
-  uint8_t* M2 = (uint8_t*)calloc((M / 2) * (N / 2), sizeof(uint8_t));
+  float* M2 = (float*)calloc((M / 2) * (N / 2), sizeof(float));
   tofree_.push_back(M2);
   {
     printf("M2\n");
@@ -240,7 +240,7 @@ inline void CPUAccelerationConvOp::coppersmith_winograd(
   }
 
   // M3 = S4*B22
-  uint8_t* M3 = (uint8_t*)calloc((M / 2) * (N / 2), sizeof(uint8_t));
+  float* M3 = (float*)calloc((M / 2) * (N / 2), sizeof(float));
   tofree_.push_back(M3);
   {
     printf("M3\n");
@@ -249,7 +249,7 @@ inline void CPUAccelerationConvOp::coppersmith_winograd(
   }
 
   // M4 = A22*T4
-  uint8_t* M4 =(uint8_t*) calloc((M / 2) * (N / 2), sizeof(uint8_t));
+  float* M4 =(float*) calloc((M / 2) * (N / 2), sizeof(float));
   tofree_.push_back(M4);
   {
     printf("M4\n");
@@ -258,7 +258,7 @@ inline void CPUAccelerationConvOp::coppersmith_winograd(
   }
 
   // M5 = S1*T1
-  uint8_t* M5 = (uint8_t*)calloc((M / 2) * (N / 2), sizeof(uint8_t));
+  float* M5 = (float*)calloc((M / 2) * (N / 2), sizeof(float));
   tofree_.push_back(M5);
   {
     printf("M5\n");
@@ -266,7 +266,7 @@ inline void CPUAccelerationConvOp::coppersmith_winograd(
   }
 
   // M6 = S2*T2
-  uint8_t* M6 = (uint8_t*)calloc((M / 2) * (N / 2), sizeof(uint8_t));
+  float* M6 = (float*)calloc((M / 2) * (N / 2), sizeof(float));
   tofree_.push_back(M6);
   {
     printf("M6\n");
@@ -274,7 +274,7 @@ inline void CPUAccelerationConvOp::coppersmith_winograd(
   }
 
   // M7 = S3*T3
-  uint8_t* M7 = (uint8_t*)calloc((M / 2) * (N / 2), sizeof(uint8_t));
+  float* M7 = (float*)calloc((M / 2) * (N / 2), sizeof(float));
   tofree_.push_back(M7);
   {
     printf("M7\n");
@@ -310,7 +310,7 @@ inline void CPUAccelerationConvOp::coppersmith_winograd(
   }
 }
 
-inline float CPUAccelerationConvOp::im2col_get_pixel(uint8_t* im, int height,
+inline float CPUAccelerationConvOp::im2col_get_pixel(float* im, int height,
                                                      int width, int channels,
                                                      int row, int col,
                                                      int channel, int pad) {
@@ -321,10 +321,10 @@ inline float CPUAccelerationConvOp::im2col_get_pixel(uint8_t* im, int height,
   return im[col + width * (row + height * channel)];
 }
 
-inline void CPUAccelerationConvOp::im2col_cpu(uint8_t* data_im, int channels,
+inline void CPUAccelerationConvOp::im2col_cpu(float* data_im, int channels,
                                               int height, int width, int ksize,
                                               int stride, int pad,
-                                              uint8_t* data_col) {
+                                              float* data_col) {
   int c, h, w;
   // 计算卷基层输出图像的高 和宽
   int height_col = (height + 2 * pad - ksize) / stride + 1;
