+++ src/ops/conv.cpp
@@ -72,10 +72,10 @@ inline void CPUConvOp::forward_compute() {
   auto input_tensor = getInputs()[0];
   auto output_tensor = getOutputs()[0];
 
-  uint8_t* input = reinterpret_cast<uint8_t*>(input_tensor->data_ptr);
-  uint8_t* output = reinterpret_cast<uint8_t*>(output_tensor->data_ptr);
-  uint8_t* weight = reinterpret_cast<uint8_t*>(weight_->data_ptr);
-  uint8_t* bias = bias_ ? reinterpret_cast<uint8_t*>(bias_->data_ptr) : nullptr;
+  float* input = reinterpret_cast<float*>(input_tensor->data_ptr);
+  float* output = reinterpret_cast<float*>(output_tensor->data_ptr);
+  float* weight = reinterpret_cast<float*>(weight_->data_ptr);
+  float* bias = bias_ ? reinterpret_cast<float*>(bias_->data_ptr) : nullptr;
 
   int ni = input_tensor->n_batch;
   int ci = input_tensor->channel;
@@ -95,13 +95,13 @@ inline void CPUConvOp::forward_compute() {
   int ph = param_.ph;
   int pw = param_.pw;
 
-  uint8_t* temp_weight = nullptr;
+  float* temp_weight = nullptr;
   int x = 0, y = 0;
   if (dh > 1 || dw > 1) {
     kh = (kh - 1) * dh + 1;
     kw = (kw - 1) * dw + 1;
     temp_weight =
-        reinterpret_cast<uint8_t*>(malloc(sizeof(uint8_t) * kw * kh * ci * co));
+        reinterpret_cast<float*>(malloc(sizeof(float) * kw * kh * ci * co));
     x = -1;
     y = -1;
     // padding
